# services/api-gateway/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SBSMA - Scalable Backend System using Microservices API",
    "description": "Comprehensive REST API Documentation for the Scalable E-Commerce Backend System composed of 5 independent microservices.\n\n### Microservices Included:\n1. **User Service** (`:3001`)\n2. **Product Service** (`:3002`)\n3. **Order Service** (`:3003`)\n4. **Payment Service** (`:3004`)\n5. **Notification Service** (`:3005`)\n\nAll endpoints are accessible via this **API Gateway** (`:3000`).",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "API Gateway Server (Port 3000)"
    }
  ],
...
  "paths": {
...
    "/api/orders": {
      "get": {
        "tags": ["Orders"],
        "summary": "Get all orders (Read - List)",
        "responses": { "200": { "description": "Orders list retrieved" } }
      },
      "post": {
        "tags": ["Orders"],
        "summary": "Create new order (Create - Interacts with Product & Notification services)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["user_id", "items", "shipping_address"],
                "properties": {
                  "user_id": { "type": "integer", "example": 1 },
                  "shipping_address": { "type": "string", "example": "42 MG Road, Bengaluru, Karnataka" },
                  "items": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "required": ["product_id", "quantity"],
                      "properties": {
                        "product_id": { "type": "integer", "example": 1 },
                        "quantity": { "type": "integer", "example": 1 }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Order created and stock reserved" },
          "400": { "description": "Insufficient stock or invalid input" },
          "404": { "description": "Product not found" }
        }
      }
    },
    "/api/orders/{id}": {
      "get": {
        "tags": ["Orders"],
        "summary": "Get order by ID with line items (Read - Single)",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": {
          "200": { "description": "Order details retrieved" },
          "404": { "description": "Order not found" }
        }
      },
      "delete": {
        "tags": ["Orders"],
        "summary": "Cancel and delete order (Delete - Restores reserved stock)",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": {
          "200": { "description": "Order cancelled and inventory restored" },
          "404": { "description": "Order not found" }
        }
      }
    },
    "/api/orders/{id}/status": {
      "put": {
        "tags": ["Orders"],
        "summary": "Update order status (Update)",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["status"],
                "properties": {
                  "status": {
                    "type": "string",
                    "enum": ["PENDING", "CONFIRMED", "SHIPPED", "DELIVERED", "CANCELLED"],
                    "example": "SHIPPED"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Order status updated successfully" },
          "400": { "description": "Invalid status value" },
          "404": { "description": "Order not found" }
        }
      }
    },
    "/api/payments": {
      "get": {
        "tags": ["Payments"],
        "summary": "Get all payment transactions (Read - List)",
        "responses": { "200": { "description": "Payment transactions retrieved" } }
      },
      "post": {
        "tags": ["Payments"],
        "summary": "Process payment for order (Create - Confirms order & notifies user)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["order_id", "user_id", "amount", "payment_method"],
                "properties": {
                  "order_id": { "type": "integer", "example": 1 },
                  "user_id": { "type": "integer", "example": 1 },
                  "amount": { "type": "number", "example": 119999.00 },
                  "currency": { "type": "string", "example": "INR" },
                  "payment_method": {
                    "type": "string",
                    "enum": ["CREDIT_CARD", "DEBIT_CARD", "UPI", "NET_BANKING", "CASH_ON_DELIVERY"],
                    "example": "UPI"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Payment processed and order confirmed" },
          "400": { "description": "Invalid payment method or missing fields" }
        }
      }
    },
    "/api/payments/{id}": {
      "get": {
        "tags": ["Payments"],
        "summary": "Get payment details by ID (Read - Single)",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": {
          "200": { "description": "Payment details retrieved" },
          "404": { "description": "Payment not found" }
        }
      },
      "put": {
        "tags": ["Payments"],
        "summary": "Update payment details/status (Update)",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": { "type": "string", "example": "COMPLETED" },
                  "payment_method": { "type": "string", "example": "CREDIT_CARD" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Payment updated successfully" },
          "404": { "description": "Payment not found" }
        }
      },
      "delete": {
        "tags": ["Payments"],
        "summary": "Cancel / Refund payment (Delete)",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": {
          "200": { "description": "Payment marked as refunded" },
          "404": { "description": "Payment not found" }
        }
      }
    },
...
  }
}